pyproject.toml
[project]
name = "fastapi-tts"
version = "0.1.0"
description = "FastAPI TTS server with ElevenLabs API"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn>=0.34,<1",
    "elevenlabs>=1.0,<2",
    "python-dotenv>=1.0,<2",
    "sounddevice",
    "numpy",
    "pynput",
    "websockets",
]
